feat: Merged real and complex vectors to one, with traits implemented based on item type
