+++ ROADMAP.md
@@ -1,0 +1,6 @@
+# Roadmap
+
+- [x] Create a common vector type for a generic number type T.
+- [ ] Implement methods for real and fixed numbers in separate submodules.
+- [ ] Implement a simplified plotting interface/example.
+- [ ] Implement readable `Format` and `Display` traits for real and complex vectors.
