--- src/fixed.rs
@@ -1,10 +1,0 @@
-//--------------------------------------------------------------//
-// Copyright 2021 Norsk Datateknikk AS                          //
-//--------------------------------------------------------------//
-// This file is subject to the terms and conditions defined in  //
-// file 'LICENSE', which is part of this source code package.   //
-//--------------------------------------------------------------//
-
-
-pub mod real;
-pub mod complex;
